requirements.txt
langchain==0.2.10
langchain_chroma==0.1.1
langchain_community==0.0.6
langchain-google-genai==1.0.8
langchain_experimental==0.0.62
sentence-transformers==2.7.0
streamlit==1.33.0
python-dotenv==1.0.1
SQLAlchemy==1.4.39
